working

+++ src/dft.cpp
@@ -1,17 +1,17 @@
 #include "dft.hpp"
-const float pi2 = 6.28318530718;
+const float pi2 = 6.28318530718f;
 
 
-void dft(float* re,float* im,uint16_t* data,size_t n)
+void dft(float* re,float* im,float* data,size_t n)
 {
-    for(int k=0;k<n/2;k++)
+    for(size_t k=0;k<n/2;k++)
     {
         re[k]=0;
         im[k]=0;
-        for(int j=0;j<n;j++)
+        for(size_t j=0;j<n;j++)
         {
-            re[k] += data[j]*cos(pi2*j*k/n);
-            im[k] -= data[j]*sin(pi2*j*k/n);
+            re[k] += data[j]*cosf(pi2*j*k/n);
+            im[k] -= data[j]*sinf(pi2*j*k/n);
         }
     }
 }
@@ -19,9 +19,9 @@ void dft(float* re,float* im,uint16_t* data,size_t n)
 void mag(float* re,float* im,float* mag,size_t n,float* maxmag)
 {
     *maxmag = 0;
-    for(int i=0;i<n;i++)
+    for(size_t i=0;i<n;i++)
     {
-        mag[i]=sqrt(re[i]*re[i]+im[i]*im[i]);
+        mag[i]=sqrtf(re[i]*re[i]+im[i]*im[i]);
         if(mag[i]>*maxmag)
             *maxmag = mag[i];
     }
