+++ src/dft.hpp
@@ -5,10 +5,9 @@
 #include <stddef.h>
 #include <math.h>
 #include <string.h>
+#include <complex.h>
 
-
-void dft(float* re,float* im,uint16_t* data,size_t n);
+void dft(float* re,float* im,float* data,size_t n);
 void mag(float* re,float* im,float* mag,size_t n,float* maxmag);
-
 
 #endif // __DFT_H__
